♻️ Refactor :  Mask, TourStep 폴더 구축 및 상대 위치를 절대 위치로 변경
