+++ src/components/data-display/Tour/Mask/Mask.scss
@@ -14,9 +14,8 @@
 }
 
 .JinniMask {
-  position: fixed;
+  position: absolute;
   top: 0;
   left: 0;
-  width: 100vw;
-  height: 100vh;
+  width: 100%;
 }
