+++ src/components/data-display/Tour/Mask/index.ts
@@ -1,0 +1,1 @@
+export { default as Mask } from './Mask';
