+++ src/components/data-display/Tour/Tour.hooks.ts
@@ -1,12 +1,6 @@
-import React, { useLayoutEffect, useContext, useEffect } from 'react';
+import { useLayoutEffect, useContext, useEffect } from 'react';
 import { TourProps } from './Tour';
-import { TourStepProps } from './TourStep';
 import TourContext from './Tour.contexts';
-import { getAnchorElCoordinate, getPopperCoordinate } from '@/utils/popper';
-import {
-  placementToAnchorOrigin,
-  placementToPopperOrigin
-} from '@/utils/popper';
 
 export const useTour = () => {
   const value = useContext(TourContext);
@@ -42,52 +36,4 @@ export const useKeydown = ({ onClose }: Pick<TourProps, 'onClose'>) => {
       document.removeEventListener('keydown', handleEscape);
     };
   }, [onClose]);
-};
-
-export const usePlacement = ({
-  tourStepElRef,
-  anchorEl,
-  placement,
-  show
-}: Required<
-  Pick<TourStepProps, 'anchorEl' | 'placement'> & {
-    tourStepElRef: React.RefObject<HTMLElement>;
-    show: boolean;
-  }
->) => {
-  useLayoutEffect(() => {
-    if (!show) return;
-
-    const setPosition = () => {
-      const tourStepEl = tourStepElRef.current;
-      if (!tourStepEl) return;
-
-      const anchorOrigin = placementToAnchorOrigin(placement);
-      const tourStepOrigin = placementToPopperOrigin(placement);
-      const anchorCoordinate = getAnchorElCoordinate({
-        anchorEl,
-        anchorOrigin
-      });
-      const tourStepElCoordinate = getPopperCoordinate({
-        anchorCoordinate,
-        popperOrigin: tourStepOrigin,
-        popperEl: tourStepEl
-      });
-
-      const { top, left } = tourStepElCoordinate;
-      tourStepEl.style.top = `${top}px`;
-      tourStepEl.style.left = `${left}px`;
-    };
-
-    setPosition();
-    window.addEventListener('resize', setPosition);
-    window.addEventListener('scroll', setPosition);
-    const observer = new ResizeObserver(setPosition);
-    observer.observe(anchorEl);
-    return () => {
-      window.removeEventListener('resize', setPosition);
-      window.removeEventListener('scroll', setPosition);
-      observer.unobserve(anchorEl);
-    };
-  }, [tourStepElRef, anchorEl, show, placement]);
 };
