+++ src/components/data-display/Tour/TourStep/TourStep.scss
@@ -40,7 +40,7 @@
 
 .JinniTourStep {
   display: inline-block;
-  position: fixed;
+  position: absolute;
 
   .JinniTourStepContent {
     position: relative;
