+++ src/components/data-display/Tour/TourStep/index.ts
@@ -1,0 +1,1 @@
+export { default as TourStep } from './TourStep';
