+++ src/components/data-display/Tour/index.ts
@@ -1,2 +1,2 @@
 export { default as Tour } from './Tour';
-export { default as TourStep } from './TourStep';
+export { TourStep } from './TourStep';
